requirements.txt
django==2.2.28
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyJWT==2.4.0
Pillow==9.0.1
requests==2.26.0
sorl-thumbnail==12.7.0